// src/server/V2/fetchData/fetch-eau/api-eau.d.ts
export interface ResultatDis {
  count: number;
  first: string | null;
  last: string | null;
  prev: string | null;
  next: string | null;
  api_version: string;
  data: Array<DataItemResultatDis>;
}

export interface DataItemResultatDis {
  code_departement: string;
  nom_departement: string;
  code_prelevement: string;
  code_parametre: string;
  code_parametre_se: string;
  code_parametre_cas: string | null;
  libelle_parametre: string;
  libelle_parametre_maj: string;
  libelle_parametre_web: string | null;
  code_type_parametre: string;
  code_lieu_analyse: string;
  resultat_alphanumerique: string;
  resultat_numerique: number;
  libelle_unite: string;
  code_unite: string;
  limite_qualite_parametre: string | null;
  reference_qualite_parametre: string | null;
  code_commune: string;
  nom_commune: string;
  nom_uge: string;
  nom_distributeur: string;
  nom_moa: string;
  date_prelevement: string;
  conclusion_conformite_prelevement: string;
  conformite_limites_bact_prelevement: string;
  conformite_limites_pc_prelevement: string;
  conformite_references_bact_prelevement: string;
  conformite_references_pc_prelevement: string;
  reference_analyse: string | null;
  code_installation_amont: string;
  nom_installation_amont: string;
  reseaux: Array<ReseauResultatDis>;
}

export interface ReseauResultatDis {
  code: string;
  nom: string;
  debit?: string;
}